requirements.txt
flask
python-dotenv
openai
aiohttp
requests
gspread
google-auth
oauth2client